Fix YGG spam

+++ pynyaata/config.py
@@ -11,9 +11,9 @@ load_dotenv()
 IS_DEBUG = environ.get('FLASK_ENV', 'production') == 'development'
 ADMIN_USERNAME = environ.get('ADMIN_USERNAME', 'admin')
 ADMIN_PASSWORD = generate_password_hash(environ.get('ADMIN_PASSWORD', 'secret'))
-APP_PORT = environ.get('FLASK_PORT', 5000)
-CACHE_TIMEOUT = environ.get('CACHE_TIMEOUT', 60 * 60)
-REQUESTS_TIMEOUT = environ.get('REQUESTS_TIMEOUT', 5)
+APP_PORT = int(environ.get('FLASK_PORT', 5000))
+CACHE_TIMEOUT = int(environ.get('CACHE_TIMEOUT', 60 * 60))
+REQUESTS_TIMEOUT = int(environ.get('REQUESTS_TIMEOUT', 5))
 BLACKLIST_WORDS = environ.get('BLACKLIST_WORDS', '').split(',') if environ.get('BLACKLIST_WORDS', '') else []
 MYSQL_ENABLED = False
